remove is_external
